add actions to vm v2

+++ pkg/subsystems/k8s/vm.go
@@ -6,13 +6,15 @@ import (
 	"github.com/google/uuid"
 	"go-deploy/pkg/subsystems/k8s/keys"
 	"go-deploy/pkg/subsystems/k8s/models"
+	"go-deploy/utils"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"log"
 	"time"
 )
 
 func (client *Client) ReadVM(id string) (*models.VmPublic, error) {
 	makeError := func(err error) error {
-		return fmt.Errorf("failed to read k8s vm. details: %w", err)
+		return fmt.Errorf("failed to read k8s vm %s. details: %w", id, err)
 	}
 
 	vm, err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).Get(context.TODO(), id, metav1.GetOptions{})
@@ -31,7 +33,7 @@ func (client *Client) ReadVM(id string) (*models.VmPublic, error) {
 
 func (client *Client) CreateVM(public *models.VmPublic) (*models.VmPublic, error) {
 	makeError := func(err error) error {
-		return fmt.Errorf("failed to create k8s vm. details: %w", err)
+		return fmt.Errorf("failed to create k8s vm %s. details: %w", public.ID, err)
 	}
 
 	vms, err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).List(context.TODO(), metav1.ListOptions{
@@ -59,10 +61,24 @@ func (client *Client) CreateVM(public *models.VmPublic) (*models.VmPublic, error
 
 func (client *Client) UpdateVM(public *models.VmPublic) (*models.VmPublic, error) {
 	makeError := func(err error) error {
-		return fmt.Errorf("failed to update k8s vm. details: %w", err)
+		return fmt.Errorf("failed to update k8s vm %s. details: %w", public.ID, err)
 	}
 
-	manifest := CreateVmManifest(public)
+	if public.ID == "" {
+		log.Println("no id supplied when updating k8s vm. assuming it was deleted")
+		return nil, nil
+	}
+
+	vm, err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).Get(context.TODO(), public.ID, metav1.GetOptions{})
+	if err != nil {
+		if IsNotFoundErr(err) {
+			return nil, nil
+		}
+
+		return nil, makeError(err)
+	}
+
+	manifest := CreateVmManifest(public, vm.ObjectMeta.ResourceVersion)
 	res, err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).Update(context.TODO(), manifest, metav1.UpdateOptions{})
 	if err != nil {
 		return nil, makeError(err)
@@ -73,7 +89,12 @@ func (client *Client) UpdateVM(public *models.VmPublic) (*models.VmPublic, error
 
 func (client *Client) DeleteVM(id string) error {
 	makeError := func(err error) error {
-		return fmt.Errorf("failed to delete k8s vm. details: %w", err)
+		return fmt.Errorf("failed to delete k8s vm %s. details: %w", id, err)
+	}
+
+	if id == "" {
+		log.Println("no id supplied when deleting k8s vm. assuming it was deleted")
+		return nil
 	}
 
 	err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).Delete(context.TODO(), id, metav1.DeleteOptions{})
@@ -83,6 +104,38 @@ func (client *Client) DeleteVM(id string) error {
 
 	err = client.waitVmDeleted(id)
 	if err != nil {
+		return makeError(err)
+	}
+
+	return nil
+}
+
+// DeleteVMIs deletes all VMIs for a VM.
+//
+// It assumes that the VirtualMachine parent has the same name as the VMIs.
+func (client *Client) DeleteVMIs(vmID string) error {
+	makeError := func(err error) error {
+		return fmt.Errorf("failed to delete k8s vmis for vm %s. details: %w", vmID, err)
+	}
+
+	if vmID == "" {
+		log.Println("no vm id supplied when deleting k8s vmis. assuming it was deleted")
+		return nil
+	}
+
+	vm, err := client.KubeVirtK8sClient.KubevirtV1().VirtualMachines(client.Namespace).Get(context.TODO(), vmID, metav1.GetOptions{})
+	if err != nil {
+		if IsNotFoundErr(err) {
+			return nil
+		}
+
+		return makeError(err)
+	}
+
+	err = client.KubeVirtK8sClient.KubevirtV1().VirtualMachineInstances(client.Namespace).Delete(context.TODO(), vm.Name, metav1.DeleteOptions{
+		GracePeriodSeconds: utils.Int64Ptr(0),
+	})
+	if err != nil && !IsNotFoundErr(err) {
 		return makeError(err)
 	}
